rework player attacks with PlayerAttackEvent
