+++ src/entities/mod.rs
@@ -1,19 +1,18 @@
 use crate::entities::entity::EntityPlugin;
 use crate::entities::event::EventPlugin;
 use crate::entities::mob::MobPlugin;
+use crate::entities::player::spells::SpellPlugin;
 use crate::entities::player::PlayerPlugin;
-use crate::entities::spells::SpellPlugin;
 use bevy::app::PluginGroupBuilder;
 use bevy::prelude::PluginGroup;
-
-pub mod entity;
-pub mod mob;
-pub mod player;
+use crate::entities::player::attacks::AttackPlugin;
+use crate::entities::player::melee::MeleePlugin;
 
 pub mod data;
-
+pub mod entity;
 pub mod event;
-mod spells;
+pub mod mob;
+pub mod player;
 
 pub struct EntityPlugins;
 
@@ -24,6 +23,8 @@ impl PluginGroup for EntityPlugins {
             .add(EntityPlugin)
             .add(MobPlugin)
             .add(PlayerPlugin)
+            .add(AttackPlugin)
+            .add(MeleePlugin)
             .add(SpellPlugin)
     }
 }
